Completed Value Added Features and Refactors

+++ HotelListing.API/Controllers/CountriesV2Controller.cs
@@ -1,22 +1,21 @@
 ﻿using AutoMapper;
 using HotelListing.API.Core.Contracts;
 using HotelListing.API.Core.Exceptions;
-using HotelListing.API.Core.Models;
 using HotelListing.API.Core.Models.Country;
-using HotelListing.Data;
 using HotelListing.Common;
+using HotelListing.Data;
 using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc;
+using Microsoft.AspNetCore.OData.Query;
 using Microsoft.EntityFrameworkCore;
-using static HotelListing.API.Core.Models.QueryParameters;
 
 namespace HotelListing.API.Controllers
 {
     [Route("api/v{version:apiVersion}/countries")]
     [ApiController]
-    [ApiVersion("1.0")]
+    [ApiVersion("2.0")]
 
-    public class CountriesController : ControllerBase
+    public class CountriesV2Controller : ControllerBase
     {
         //Needs to be refactored from boilerplate
         //1. Don't want to call db context directly from controller...Create Repositories
@@ -24,9 +23,9 @@ namespace HotelListing.API.Controllers
 
         private readonly ICountriesRepository _countriesRepository;
         private readonly IMapper _mapper;
-        private readonly ILogger<CountriesController> _logger;
+        private readonly ILogger<CountriesV2Controller> _logger;
 
-        public CountriesController(ICountriesRepository countriesRepository, IMapper mapper, ILogger<CountriesController> logger)
+        public CountriesV2Controller(ICountriesRepository countriesRepository, IMapper mapper, ILogger<CountriesV2Controller> logger)
         {
             _countriesRepository = countriesRepository;
             _mapper = mapper;
@@ -34,20 +33,14 @@ namespace HotelListing.API.Controllers
         }
 
         // GET: api/Countries
-        [HttpGet("GetAll")]
+        [HttpGet]
+        [EnableQuery]
         public async Task<ActionResult<IEnumerable<GetCountryDTO>>> GetCountries()
         {
-            var countries = await _countriesRepository.GetAllAsync<GetCountryDTO>();
-            return Ok(countries);
-
-        }
-
-        // GET: api/Countries/?StartIndex=0&pageSize25&PageNumber=1
-        [HttpGet]
-        public async Task<ActionResult<PagedResult<GetCountryDTO>>> GetPagedCountries([FromQuery] QueryParameters queryParameters)
-        {
-            var pagedCountriesResult = await _countriesRepository.GetAllAsync<GetCountryDTO>(queryParameters);
-            return Ok(pagedCountriesResult);
+            var countries = await _countriesRepository.GetAllAsync();
+            var records = _mapper.Map<List<GetCountryDTO>>(countries);
+            //Select * from Countries
+            return Ok(records);
 
         }
 
@@ -56,7 +49,15 @@ namespace HotelListing.API.Controllers
         public async Task<ActionResult<CountryDTO>> GetCountry(int id)
         {
             var country = await _countriesRepository.GetDetails(id);
-            return Ok(country);
+
+            if (country == null)
+            {
+                throw new NotFoundException(nameof(GetCountry), id);
+            }
+
+            var results = _mapper.Map<CountryDTO>(country);
+
+            return Ok(results);
         }
 
         // PUT: api/Countries/5
@@ -65,9 +66,25 @@ namespace HotelListing.API.Controllers
         [Authorize]
         public async Task<IActionResult> PutCountry(int id, UpdateCountryDTO updateCountryDTO)
         {
+            if (id != updateCountryDTO.Id)
+            {
+                return BadRequest("Invalid Record Id");
+            }
+
+            //_context.Entry(country).State = EntityState.Modified;
+
+            var country = await _countriesRepository.GetAsync(id); //Finding the country from the database using ID..it's being "tracked"
+
+            if (country == null)
+            {
+                throw new NotFoundException(nameof(GetCountries), id);
+            }
+
+            _mapper.Map(updateCountryDTO, country); //Automatically told entity framework that I have changed this to modified and automatically assigned values from the right left to the right side
+
             try
             {
-                await _countriesRepository.UpdateAsync(id, updateCountryDTO);
+                await _countriesRepository.UpdateAsync(country);
             }
             catch (DbUpdateConcurrencyException)
             {
@@ -81,7 +98,7 @@ namespace HotelListing.API.Controllers
                 }
             }
 
-            return NoContent();
+            return NoContent();//This is a good response 204 No Content
         }
 
         // POST: api/Countries
@@ -90,9 +107,9 @@ namespace HotelListing.API.Controllers
         [Authorize]
         public async Task<ActionResult<CountryDTO>> PostCountry(CreateCountryDTO createCountryDTO)
         {
-            var country = await _countriesRepository.AddAsync<CreateCountryDTO, CountryDTO>(createCountryDTO);
-            return CreatedAtAction(nameof(GetCountry), new { id = country.Id }, country);
 
+            var country = await _countriesRepository.AddAsync<CreateCountryDTO, CountryDTO>(createCountryDTO);
+            return CreatedAtAction(nameof(GetCountry), new {id = country.Id }, country);
             //After that is done you can get there by calling "GetCountry" url endpoint by the new id(country.id) from the country object
         }
 
@@ -101,8 +118,15 @@ namespace HotelListing.API.Controllers
         [Authorize(Roles = SD.Role_Admin)]
         public async Task<IActionResult> DeleteCountry(int id)
         {
+            var country = await _countriesRepository.GetAsync(id);
+            if (country == null)
+            {
+                throw new NotFoundException(nameof(GetCountries), id);
+            }
+
             await _countriesRepository.DeleteAsync(id);
-            return NoContent();
+
+            return NoContent();//204 No Content
         }
 
         private async Task<bool> CountryExists(int id)
